pyproject.toml
[project]
name = "wedaa-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.3",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.2",
    "mysql-connector-python>=9.3.0",
    "openai>=1.77.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "pymongo>=4.12.1",
    "pymysql>=1.1.1",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.40",
]
